+++ src/components/mainpage/sidebar/AdminDrawer/handleToggle.ts
@@ -1,0 +1,5 @@
+import type { Dispatch, SetStateAction } from "react"
+
+export function handleToggle(setter: Dispatch<SetStateAction<boolean>>) {
+    setter(prev => !prev)
+}
